balena-containers/raspberrypirfid/requirements.txt
paho-mqtt
RPi.GPIO
mfrc522